requirements.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.1.0
attrs==22.1.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
bqplot==0.12.36
branca==0.6.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
cftime==1.6.2
charset-normalizer==2.1.1
click==8.1.3
climateserv==0.0.24
colorama==0.4.6
colour==0.1.5
contourpy==1.0.6
cryptography==38.0.3
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.3
django-allauth==0.51.0
django-import-export==3.0.1
docopt==0.6.2
earthengine-api==0.1.331
ee-extra==0.0.14
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.16.2
ffmpeg-python==0.2.0
filelock==3.8.0
folium==0.13.0
fonttools==4.38.0
future==0.18.2
gdown==4.5.3
geeadd==0.5.6
geemap==0.17.2
geocoder==1.38.1
geojson==2.5.0
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
httplib2==0.21.0
idna==3.4
ipyevents==2.0.1
ipyfilechooser==0.6.0
ipykernel==6.17.1
ipyleaflet==0.17.2
ipython==8.6.0
ipython-genutils==0.2.0
ipytree==0.2.2
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
json5==0.9.10
jsonschema==4.17.0
jupyter-server==1.23.2
jupyter_client==7.4.5
jupyter_core==5.0.0
jupyterlab==3.5.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
jupyterlab_server==2.16.3
kiwisolver==1.4.4
logzero==1.7.0
MarkupPy==1.14
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
netCDF4==1.6.1
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.4
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.4
plotly==5.11.0
prometheus-client==0.15.0
prompt-toolkit==3.0.32
protobuf==4.21.9
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyCRS==1.0.2
Pygments==2.13.0
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.2
pyshp==2.3.1
PySocks==1.7.1
python-box==6.1.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.6
pywin32==305
pywinpty==2.0.9
PyYAML==6.0
pyzmq==24.0.1
ratelim==0.1.6
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
sankee==0.2.0
scooby==0.7.0
Send2Trash==1.8.0
Shapely==1.8.5.post1
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.6.1
tablib==3.2.1
tenacity==8.1.0
terminado==0.17.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
traittypes==0.2.1
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
whitebox==2.2.0
whiteboxgui==2.2.0
widgetsnbextension==4.0.3
xlrd==2.0.1
xlwt==1.3.0
xyzservices==2022.9.0
yarg==0.1.9
